# jac-cloud/jac_cloud/tests/openapi_specs.yaml
components:
  schemas:
...
    Expiration:
      description: Key Expiration.
      properties:
        count:
          default: 60
          exclusiveMinimum: 0.0
          title: Count
          type: integer
        interval:
          default: days
          enum:
          - seconds
          - minutes
          - hours
          - days
          title: Interval
          type: string
      title: Expiration
      type: object
    GenerateKey:
      description: User Creation Request Model.
      properties:
        expiration:
          $ref: '#/components/schemas/Expiration'
        name:
          minLength: 1
          title: Name
          type: string
        nodes:
          items:
            type: string
          title: Nodes
          type: array
        walkers:
          items:
            type: string
          title: Walkers
          type: array
      required:
      - name
      title: GenerateKey
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    KeyIDs:
      description: User Creation Request Model.
      properties:
        ids:
          items:
            type: string
          minItems: 1
          title: Ids
          type: array
      required:
      - ids
      title: KeyIDs
      type: object
    UserChangePassword:
      description: User Creation Request Model.
      properties:
        new_password:
          title: New Password
          type: string
        old_password:
          title: Old Password
          type: string
      required:
      - old_password
      - new_password
      title: UserChangePassword
      type: object
    UserForgotPassword:
      description: User Creation Request Model.
      properties:
        email:
          format: email
          title: Email
          type: string
      required:
      - email
      title: UserForgotPassword
      type: object
    UserRegister:
      properties:
        email:
          format: email
          title: Email
          type: string
        name:
          title: Name
          type: string
        password:
          title: Password
          type: string
      required:
      - email
      - password
      - name
      title: UserRegister
      type: object
    UserRequest:
      description: User Creation Request Model.
      properties:
        email:
          format: email
          title: Email
          type: string
        password:
          title: Password
          type: string
      required:
      - email
      - password
      title: UserRequest
      type: object
    UserResetPassword:
      description: User Creation Request Model.
      properties:
        code:
          title: Code
          type: string
        password:
          title: Password
          type: string
      required:
      - code
      - password
      title: UserResetPassword
      type: object
    UserVerification:
      description: User Verification Request Model.
      properties:
        code:
          title: Code
          type: string
      required:
      - code
      title: UserVerification
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    APIKeyHeader:
      in: header
      name: test_key
      type: apiKey
    APIKeyQuery:
      in: query
      name: test_key
      type: apiKey
    HTTPBearer:
      scheme: bearer
      type: http
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /user/change_password:
    post:
      description: Register user API.
      operationId: change_password_user_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserChangePassword'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Change Password
      tags:
      - User APIs
  /user/forgot_password:
    post:
      description: Forgot password API.
      operationId: forgot_password_user_forgot_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserForgotPassword'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Forgot Password
      tags:
      - User APIs
  /user/login:
    post:
      description: Login user API.
      operationId: login_user_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - User APIs
  /user/register:
    post:
      description: Register user API.
      operationId: register_user_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
      tags:
      - User APIs
  /user/reset_password:
    post:
      description: Reset password API.
      operationId: reset_password_user_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserResetPassword'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reset Password
      tags:
      - User APIs
  /user/send-verification-code:
    post:
      description: Verify user API.
      operationId: send_verification_code_user_send_verification_code_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Send Verification Code
      tags:
      - User APIs
  /user/verify:
    post:
      description: Verify user API.
      operationId: verify_user_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserVerification'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Verify
      tags:
      - User APIs
...
  /webhook:
    get:
      description: Get keys API.
      operationId: get_webhook_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get
      tags:
      - Webhook APIs
  /webhook/delete:
    delete:
      description: Delete keys API.
      operationId: delete_webhook_delete_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyIDs'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete
      tags:
      - Webhook APIs
  /webhook/extend/{id}:
    patch:
      description: Generate key API.
      operationId: extend_webhook_extend__id__patch
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Expiration'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Extend
      tags:
      - Webhook APIs
  /webhook/generate-key:
    post:
      description: Generate key API.
      operationId: generate_key_webhook_generate_key_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateKey'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Generate Key
      tags:
      - Webhook APIs
...